Make idSkipList work again.
The problem was that basically nothing in the passed in attrSkipList could
ever get evaluated properly unless they're globals, because the environment
local to the eval() statement didn't contain much at all.  The solution is
to add the local environment from the frame where the traceback happened to
the environment the eval() runs in.