Store and then write out the string representation of the traceback and object dump
We passed the filename of the file containing the traceback and object dump to
libreport, and MainExceptionWindow. But that meant libreport and MainExceptionWindow
had to read the file and save the content to its internal structure. We need to
write traceback and object dump to a file, but we can pass its contents as string
directly to libreport and MainExceptionWindow.

There is a related patch for libreport removing the glue reading the file contents.