Reword the 'Debug' button warning (#948256)
Getting to tty1 happens only in Anaconda running a non-live installation, so we
should just warn user it may happen. Also the GUI is now on tty7, but in any
case there is no point in getting back to tty7 manually, because until the
debugging session is finished the GUI is frozen.