Introduce support for Python 3 while keeping Python 2 working
This commit introduces support for Python 3. Some of the behaviour is slightly
changed, however, it should work in general.

Libreport is still not working with Python 3, but that should be the only
blocking thing now.

Makefile and setup.py might need some modifications as well, currently only
make test3 has been added.