Add option to remove the 'Debug' option/button for mainExceptionWindow
It is neccesary to pass a function (which decides whether to allow the debug option) as a keyword argument 'allowDebug'. Default function is sys.stdout.isatty.

Default is the sys.stdout.isatty function which causes the debug option is not present when there is no terminal/tty to display the debugger.

Signed-off-by: Vojtech Trefny <vtrefny@redhat.com>