Make sure the Meh window has focus even if a dialog is displayed
Make sure that the Meh window is in focus even if the application
that crashed was displaying a dialog at the time of the crash.

This is quite an issue as both Anaconda and Blivet GUI make use of
dialogs and thus crash reporting via Python Meh could be sometimes
prevented by the Meh window not getting focus.