Drop Python 2 support
No one is apparently using the python2 subpackages and the end of
upstream support for Python 2 is in sight, so drop the python2
subpackages and Python 2 support in the codebase.

Dropping Python 2 support is pretty trivial and basically amounts
to stop using six and fixing the few places where it is used.