Spec: Move TUI back to main package
The 0.46-1 package build I think inadvertently (in the name of
simplifying the spec) moved the TUI into the -gui subpackage.
This breaks Fedora composes, as anaconda's text mode will not
start without python3-meh-gui installed, and that package is
not currently installed to the compose environment.

This commit reverts the changes to %files which broke this.