Fix handling RPM headers with RPM 4.15 and Python 3 (#1764642)
RPM 4.15 "fixed" things so these header strings are returned as
str instances in Python 3, not basestrings. Unfortunately that
breaks us, because we were using `decode()` to treat them as
Python 3-style unicode strings on both Python 2 and Python 3,
but Python 3 str instances don't have a `decode` method.

This should work for all cases (both Python 2 and Python 3, and
RPM before and after 4.15).

Signed-off-by: Adam Williamson <awilliam@redhat.com>