dump: Use exception's type and value when calculating hash
This has caused problems with later versions of Anaconda
as the traceback always leads to the same handler and
without the exception type and value there is no  distinction
between the errors.

I do not think that including the type is problematic
at all.

Including the value can result in some duplicates not
being detected as the message can contain anything.
But not including it will cause much more false
duplicates, so I think it is better to include it.

Resolves: rhbz#1848493